requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
geoalchemy2
pydantic-settings
python-dotenv
greenlet